+++ src/ormate/__init__.py
@@ -1,4 +1,5 @@
 from .adapters import SQLAlchemyAdapter, SQLModelAdapter, StorageAdapter
+from .query import QueryExpression, and_, eq, gt, gte, in_, lt, lte, ne, not_, not_in, or_
 from .repository import ModelRepository
 from .sqlalchemy.database import AsyncDatabase, Database
 
@@ -6,7 +7,19 @@ __all__ = [
     "AsyncDatabase",
     "Database",
     "ModelRepository",
+    "QueryExpression",
     "SQLAlchemyAdapter",
     "SQLModelAdapter",
     "StorageAdapter",
+    "and_",
+    "eq",
+    "gt",
+    "gte",
+    "in_",
+    "lt",
+    "lte",
+    "ne",
+    "not_",
+    "not_in",
+    "or_",
 ]